update version checker
