+++ physiofit/__main__.py
@@ -1,17 +1,36 @@
 from subprocess import run
 import sys
-import pathlib
+from pathlib import Path
+
+from threading import Thread
+import requests
 
 import physiofit
 import physiofit.ui.cli
 
+
+def get_last_version():
+    """Get last Physiofit version."""
+    try:
+        pf_path = Path(physiofit.__file__).parent
+        # Get the version from pypi
+        response = requests.get(f'https://pypi.org/pypi/physiofit/json')
+        latest_version = response.json()['info']['version']
+        with open(str(Path(pf_path, "last_version.txt")), "w") as f:
+            f.write(latest_version)
+    except:
+        pass
+
 def main():
     """The main routine"""
+
+    thread = Thread(target=get_last_version)
+    thread.start()
 
     if len(sys.argv) > 1:
         physiofit.ui.cli.main()
     else:
-        path_to_app = pathlib.Path(physiofit.__path__[0])
+        path_to_app = Path(physiofit.__file__).parent
         path_to_app = path_to_app / "ui/gui.py"
         run(["streamlit", "run", str(path_to_app)])
